move transports files into the separate directory
